requirements.txt
numpy
pillow
json
h5py
opencv-python
